refactor folder structure of board page
