+++ src/pages/Board/components/ButtonContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonContainer } from './ButtonContainer';
