+++ src/pages/Board/components/List/AddCard/index.ts
@@ -1,0 +1,1 @@
+export { default as AddCard } from './AddCard';
