+++ src/pages/Board/components/List/index.ts
@@ -1,1 +1,1 @@
-export { default as Column } from './Column';
+export { default as List } from './List';
