+++ src/pages/Board/components/index.ts
@@ -1,4 +1,4 @@
-export { Column } from './Column';
+export { List } from './List';
 export { SecondaryNavbar } from './SecondaryNavbar';
 export { default as initialData } from './fakeBoardData';
 export { AddList } from './AddList';
