+++ src/theme/themes.ts
@@ -18,7 +18,9 @@ export const defaultTheme = {
     neutral: {
       100: '#ffffff',
       200: '#f4f5f7',
+      250: '#ebecf0',
       300: '#e1e1e1',
+      350: '#d3d3d3',
       400: '#737581',
       500: '#6b788d',
       600: '#4a4b53',
@@ -50,11 +52,13 @@ export const defaultTheme = {
       800: 'hsla(0, 0%, 100%, 0.8)',
     },
     darkness: {
+      50: 'hsla(0, 0%, 0%, 0.08)',
       100: 'hsla(0, 0%, 0%, 0.1)',
       200: 'hsla(0, 0%, 0%, 0.2)',
       300: 'hsla(0, 0%, 0%, 0.3)',
       400: 'hsla(0, 0%, 0%, 0.4)',
       500: 'hsla(0, 0%, 0%, 0.5)',
     },
+    boxShadow: 'rgb(9 30 66 / 25%)',
   },
 };
